Change JTidy description: no build from source, but overrule
It is not necessary anymore to download the JTidy source code, to patch
and build it, because now the patched JTidy classes are part of the
repository. As long as the JTidy library is in the classpath *behind*
the patches during build and runtime, everything is fine. The original
classes will be overruled by the patched ones, and we can just remove
the patches as soon as the fixes are accepted and published upstream.