Extend PreJTidyFilter to also fix TOC title
Previously regex TITLE_CHAPTER removed "Ruby on Rails 2" prefix not
only in chapters, but also in TOC, because it was followed by a dash
character. This workaround removes the dash, thus leaving the prefix
untouched later during replacement in method initialiseTitle.

This fixes the subsequent error that the title was not recognised
correctly in method createIndexLink and thus the exception for the Ruby
book (no creation of link to stichwort.htm because it does not exist)
was not triggered. Now the link is *not* created anymore, as intended.