Experimentally add JsoupFilter + replace whole filter chain with it
It looks as if jsoup (http://jsoup.org) is what I have been looking for
from the very beginning: a tool which
  - can parse all openbook HTML without pre-filtering or JTidy hacking,
  - has a powerful, concise CSS/JQuery-like selector syntax (no XPath),
  - permits for easy DOM manipulation,
  - is fast,
  - outputs clean, pretty-printed HTML.

Some testing and cleaning up still needs to be done, but for the first
try things are looking pretty nice. I needed to do some logical changes
because jsoup's CSS-like selectors are often equally, but in a few cases
less powerful than XOM's XPath-like ones. Especially the section
removing the feedback forms was changed from a selector-based bulk
removal to a more procedural approach because DOM sibling selection is
not powerful as necessary.