Add member 'title' to Book. Use it in FilterChain and index.htm*.
TOC pages (index.htm*) require special handling for HTML TITLE tags
because there the book title should not been cut off to leave behind a
chapter/page title, but we explicitly want the book title as the page
title there. Instead of specifying a lot of exceptions to clean this up,
I decided to preset the book titles in public member Book.title.

FilterChain as well as BasicFilter and all its subclasses have been
updated to use the new member. In XOMUnclutterFilter.initialiseTitle it
is now used directly if(isTOCFile).