jsoup: removeFeedbackForm() now handles bogus dreamweaver_8 layout
Two pages in "dreamweaver_8" book (02_kap02_002.htm, 04_kap04_002.htm)
cannot be parsed correctly by jsoup, i.e. the feedback form is garbled
in the former and not a following sibling of an HR tag in the latter.
The jsoup mailing list has been notified about this fact. HTML tidy can
parse the documents and return the correct structure, so this might be a
jsoup problem.

This version of JsoupFiler.removeFeedbackForm() contains two versions of
"feedback form + preceding junk" removal, so as to document the changes
in both versions and how they can co-exist. Actually the new way using a
jsoup NodeVisitor probably works for bogus as well as correct DOM
structure, but it might be considerable slower for the normal case. This
is yet to be tested. If there is no significant performance penalty, the
old way of traversal might get removed until jsoup gets improved/fixed.