Remove SimpleLogger.IndentMode + all its usages again
Now that LoggingAspect has been refactored to have an extra advice
handling input, we do not need the convenience methods doing both
logging and in-/dedentation in one call anymore, because there were just
two places in JsoupFilter.initialiseTitle left which used the methods at
all.

Speaking to myself: The methods might be needed again in the future in
case of reactivation of the filter chain with more classes. Currently I
am satisfied with JsoupFilter alone, though.