LoggingAspect: get rid of indentation pointcut/advice & duplicate code
The new logging strategy is complex, but it ensures simple extendability
because now each logging advice is simply annotated by '@Log' and sets
an aspect instance variable 'message' like this:

    @Log void around() : cleanBook() {
        message = "Filtering"; proceed();
    }

If non-default logging options are needed, they can be set like this:

    @Log(logDone = false, indent = false, type = LogType.DEBUG)

The new logging mechanism works like this:
  - Advice 'after() : set(String LoggingAspect.message)' is triggered
    whenever a log advice sets the log message. It then pushes the
    message on a stack so it can be accessed later in other advice after
    a (possibly recursive) call chain with more log outputs and finally
    print "<LogMessage> - done". Without a stack, the message member
    would be overwritten by recursive logging calls. After the message
    has been pushed onto the stack, it gets printed and finally the
    indentation level is increased. Now we are all set for proceeding
    with the main application.
  - Advice 'after() : @annotation(Log)' is triggered after a loggin
    advice terminates, i.e. after the main application's 'proceed()' has
    finished. Now the indentation level is decreased again, the message
    popped from the stack and optionally printed, if '@Log.logDone()' is
    true.

PLEASE NOTE: This implementation is not thread-safe yet because
LoggingAspect is a singleton, but in multi-threaded mode running filters
might modify the log message stack concurrently. The solution is to
change the aspect's instantiation mode. I will do this in a separate
commit so as to document it clearly.