LoggingAspect: make changed from previous commit thread-safe
The previous implementation was not thread-safe yet because
LoggingAspect was a singleton, but in multi-threaded mode running
filters might modify the log message stack concurrently. The solution is
to change the aspect's instantiation mode to

    percflow(execution(* OpenbookCleaner.main(..)) || runFilter())

I.e. we create a new instance (thus also a new stack) each time a filter
runs (possibly in a new thread, depending on CLI parameter
'-t|--threading'). For the main thread we also need an instance, of
course.