LoggingAspect: use ThreadLocal variables + singleton aspect
This is an alternative to the previously used 'percflow' instantiation,
ensuring thread safety via ThreadLocal members 'message' and
'messageStack'. This makes the first pointcut and access to the two
members a bit more complicated, but has the advantage of not being
implicitly dependent on certain method names ('OpenbookCleaner.main',
'BasicFilter.run') for aspect instantiation, because we now have a
singleton aspect again.