Make book list configurable
Class Book used to be an enum with hard-coded items. Now it is a simple
class containing String values which are read from an XML configuration
file. In order to make this work in the IDE and as well as in JAR files,
the file config.xml (which was created and is parsed by XStream) is
searched in three places in this order:
  1.) ./config.xml
      If this local config file is found in either scenario (IDE or JAR)
      it will be preferred to other locations, which enables the user to
      override the default version.
  2.) resource/config.xml
      This is where the default file is located and found in Eclipse if
      no local config file exists. From there it will be maintained in
      Git and, because 'resource' is configured as an Eclipse source
      folder, added to runnable JARs when creating them from the IDE.
  3.) /config.xml
      This is where the default file is located and found in the JAR if
      no local config file exists. From there it will be read as a
      resource via Class.getResourceAsStream.

Closes #5.