Book: write character encoding iso-8859-15 to config.xml
This simplifies reading the config file from the input stream because
Xstream will respect the encoding found at the beginning of the stream.

Attention: XStream still writes its output in with UNIX line breaks
witout regard of the Java system property 'line.separator'. This is
kinda strange, but okay as long as GOC users do try to edit the config
file with a bad editor like Windows Notepad.