Options: add "-w|--write-config" for writing config.xml
This is an easy way for JAR users to extract the embedded default
config.xml to disk and edit it subsequently, e.g. in order to add new
books not contained in the upstream version. This way they do not have
to wait for me to update the Git repo and the downloadable JAR.

User self-service. :-) Refers to #5.