New class AvailableBooksChecker compares predefined to online books
This tool class has a 'main' method to be used by advanced users knowing
how to interpret the output. Basically two sets of book URLs are
compared: the list of known books from config.xml to the lists of books
found online on the Galileo Computing and Galileo Design web sites. For
each set the orphans, i.e. elements not found in the respective other
set, are determined and printed. This helps to discover new, renamed or
deleted books. It also helps to spot known books which are still
available for download, but no longer listed on the web site.