Add options '-c|--check-avail' and '-m|--check-md5'
Because Maven packages the application into a One-JAR, it is only
possible to use the main class (OpenbookCleaner) from outside, as tests
have shown. Thus, we need CLI options to dispatch to the main methods of
AvailableBooksChecker and DownloadChecker, respectively. If present, the
options will cause both helper classes to be executed in this order, but
always before OpenbookCleaner. So theoretically, users can run the
application in all three modes at once: the two diagnostic ones and the
normal book clean-up mode. Strange, but possible.