Enable GOC to parse (offline copies of) Rheinwerk online books
Automatically determine the character set via jsoup instead of setting a
fixed charset, because the "Java-Insel" book for Java 8 uses UTF-8
instead of ISO-8859-1 or windows-1252. The output stream's charset is
adjusted accordingly *after* the document was parsed.

Recognise titles in meta tags and not just in explicit title tags.

Recognise "Rheinwerk" in addition to "Galileo".