Avoid pretty-printing for DIV.listing elements containing non-PRE children
Fixes #36 and is an alternative to PR #37. It is less efficient but
combines both pretty-printing with the necessary exceptions.

Background: Currently there is no easy way to locally override HTML
pretty-printing in jsoup for certain tags, there is only a
document-global setting. Thus, the only way to partially un-apply
pretty-printing is the following:

  1. Get + preserve pretty-printed content
  2. Get + preserve content in original formatting
  3. Assign pretty-printed content to document + avoid further
     pretty-printing
  4. Replace falsely pretty-printed content by preserved original
     content

Conclusion: This is ugly because the same document which has already
been parsed + cleaned up before needs to be re-parsed two more times,
i.e. we have a total of 3 Jsoup.parse(..) calls. But - it works. :-/